Multi-threading on the server and random PSN switch
1. The server side is multi-threaded to handle continuous connection
from back-to-back spawned client processes. This makes the testing
easier because the server doesn't quit after the specified number of
ops. Otherwise, the clients and the servers have to use barriers for
synchronized testing iterations.

2. Random PSN is supported to be switched on by using -a option. If not
used, the PSN on each side is always 0. It is proved in a load testing
that even if the connections reuse the value 0, the system is still
good. Not sure in concurrent clients cases.